server/requirements.txt
# Requirements compatible with Python 3.11 and 3.13+
# Tested with Python 3.11.0+ and 3.13.0+

fastapi>=0.115.0
uvicorn[standard]>=0.32.0
httpx>=0.27.0
pydantic>=2.9.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0,<3.0.0
